refactor: move logic from controllers to services
