+++ TonPrediction.Application/Services/IPredictionService.cs
@@ -1,0 +1,34 @@
+using QYQ.Base.Common.IOCExtensions;
+using TonPrediction.Application.Output;
+
+namespace TonPrediction.Application.Services;
+
+/// <summary>
+/// 预测记录查询业务接口。
+/// </summary>
+public interface IPredictionService : ITransientDependency
+{
+    /// <summary>
+    /// 分页获取指定地址的下注记录。
+    /// </summary>
+    /// <param name="address">用户地址。</param>
+    /// <param name="status">记录状态：all/claimed/unclaimed。</param>
+    /// <param name="page">页码。</param>
+    /// <param name="pageSize">每页条数。</param>
+    /// <param name="ct">取消任务标记。</param>
+    /// <returns>下注记录列表。</returns>
+    Task<List<BetRecordOutput>> GetRecordsAsync(
+        string address,
+        string status = "all",
+        int page = 1,
+        int pageSize = 10,
+        CancellationToken ct = default);
+
+    /// <summary>
+    /// 获取指定地址的盈亏汇总。
+    /// </summary>
+    /// <param name="address">用户地址。</param>
+    /// <param name="ct">取消任务标记。</param>
+    /// <returns>盈亏信息。</returns>
+    Task<PnlOutput> GetPnlAsync(string address, CancellationToken ct = default);
+}
